created a theme switch in homepage, uniform button color
