--- src/components/Navbar.js
@@ -1,22 +1,0 @@
-import React from "react";
-import { AiOutlineMenu } from "react-icons/ai";
-import Sidebar from "./Sidebar";
-const Navbar = () => {
-  return (
-    <>
-      <nav className="border-gray-200 bg-gray-50">
-        <div className="lg:hidden flex flex-wrap items-center space-x-2 mx-auto p-4">
-          <button type="button">
-            <AiOutlineMenu className="w-6 h-6" />
-          </button>
-          <p className="self-center text-2xl font-semibold whitespace-nowrap dark:text-white">
-            STT
-          </p>
-        </div>
-      </nav>
-      <Sidebar />
-    </>
-  );
-};
-
-export default Navbar;
